+++ instrumentation/opentelemetry-instrumentation-tornado/tests/test_instrumentation.py
@@ -143,8 +143,9 @@ class TestTornadoInstrumentation(TornadoTest):
                 SpanAttributes.HTTP_HOST: "127.0.0.1:"
                 + str(self.get_http_port()),
                 SpanAttributes.HTTP_TARGET: "/",
-                SpanAttributes.NET_PEER_IP: "127.0.0.1",
+                SpanAttributes.HTTP_CLIENT_IP: "127.0.0.1",
                 SpanAttributes.HTTP_STATUS_CODE: 201,
+                "tornado.handler": "tests.tornado_test_app.MainHandler",
             },
         )
 
@@ -220,7 +221,7 @@ class TestTornadoInstrumentation(TornadoTest):
                 SpanAttributes.HTTP_HOST: "127.0.0.1:"
                 + str(self.get_http_port()),
                 SpanAttributes.HTTP_TARGET: url,
-                SpanAttributes.NET_PEER_IP: "127.0.0.1",
+                SpanAttributes.HTTP_CLIENT_IP: "127.0.0.1",
                 SpanAttributes.HTTP_STATUS_CODE: 201,
             },
         )
@@ -258,8 +259,9 @@ class TestTornadoInstrumentation(TornadoTest):
                 SpanAttributes.HTTP_HOST: "127.0.0.1:"
                 + str(self.get_http_port()),
                 SpanAttributes.HTTP_TARGET: "/error",
-                SpanAttributes.NET_PEER_IP: "127.0.0.1",
+                SpanAttributes.HTTP_CLIENT_IP: "127.0.0.1",
                 SpanAttributes.HTTP_STATUS_CODE: 500,
+                "tornado.handler": "tests.tornado_test_app.BadHandler",
             },
         )
 
@@ -292,8 +294,9 @@ class TestTornadoInstrumentation(TornadoTest):
                 SpanAttributes.HTTP_HOST: "127.0.0.1:"
                 + str(self.get_http_port()),
                 SpanAttributes.HTTP_TARGET: "/missing-url",
-                SpanAttributes.NET_PEER_IP: "127.0.0.1",
+                SpanAttributes.HTTP_CLIENT_IP: "127.0.0.1",
                 SpanAttributes.HTTP_STATUS_CODE: 404,
+                "tornado.handler": "tornado.web.ErrorHandler",
             },
         )
 
@@ -336,8 +339,9 @@ class TestTornadoInstrumentation(TornadoTest):
                 SpanAttributes.HTTP_HOST: "127.0.0.1:"
                 + str(self.get_http_port()),
                 SpanAttributes.HTTP_TARGET: "/dyna",
-                SpanAttributes.NET_PEER_IP: "127.0.0.1",
+                SpanAttributes.HTTP_CLIENT_IP: "127.0.0.1",
                 SpanAttributes.HTTP_STATUS_CODE: 202,
+                "tornado.handler": "tests.tornado_test_app.DynamicHandler",
             },
         )
 
@@ -377,8 +381,9 @@ class TestTornadoInstrumentation(TornadoTest):
                 SpanAttributes.HTTP_HOST: "127.0.0.1:"
                 + str(self.get_http_port()),
                 SpanAttributes.HTTP_TARGET: "/on_finish",
-                SpanAttributes.NET_PEER_IP: "127.0.0.1",
+                SpanAttributes.HTTP_CLIENT_IP: "127.0.0.1",
                 SpanAttributes.HTTP_STATUS_CODE: 200,
+                "tornado.handler": "tests.tornado_test_app.FinishedHandler",
             },
         )
 
@@ -481,6 +486,30 @@ class TestTornadoInstrumentation(TornadoTest):
         )
 
         self.memory_exporter.clear()
+
+
+class TestTornadoInstrumentationWithXHeaders(TornadoTest):
+    def get_httpserver_options(self):
+        return {"xheaders": True}
+
+    def test_xheaders(self):
+        response = self.fetch("/", headers={"X-Forwarded-For": "12.34.56.78"})
+        self.assertEqual(response.code, 201)
+        spans = self.get_finished_spans()
+        self.assertSpanHasAttributes(
+            spans.by_name("MainHandler.get"),
+            {
+                SpanAttributes.HTTP_METHOD: "GET",
+                SpanAttributes.HTTP_SCHEME: "http",
+                SpanAttributes.HTTP_HOST: "127.0.0.1:"
+                + str(self.get_http_port()),
+                SpanAttributes.HTTP_TARGET: "/",
+                SpanAttributes.HTTP_CLIENT_IP: "12.34.56.78",
+                SpanAttributes.HTTP_STATUS_CODE: 201,
+                SpanAttributes.NET_PEER_IP: "127.0.0.1",
+                "tornado.handler": "tests.tornado_test_app.MainHandler",
+            },
+        )
 
 
 class TornadoHookTest(TornadoTest):